requirements.txt
numpy>=1.26.4
pandas==2.2.2
scikit-learn==1.4.2
tensorflow
keras-tunernum